Make Chunker.Next read "generic" and independent from digest size.
These are both pre-works for adding write compression support to the
remote-apis-sdks.

Chunker.Next should be "generic" as it allows for a drop in implementation
of a reader that compresses on the fly. I still left the special case of
caching data in memory as it saves the effort from memory copies.

Making the chunker reads work independently from the digest size is also
useful as it allows us to not have to pre-compute the digest of the
compressed blobs. The current draft of the RE API never requires the
digest of the compressed blob at any point, and this saves us the
trouble of needing to read the data twice.

Notice that this implies that the chunker won't be actually matching
data against the supplied digest at any moment. The digest is now purely
information storage, rather than necessary for chunker logic.

As a caveat, for simplicity, I made it that we only cache in memory
files that are smaller than the *chunk* size rather than the IO buffer
size.