Revamp loadFiles for performance.
This gets rid of a few bottlenecks:

1. Recursion -> iteration. I'm not sure how much of a bottleneck this is,
but in practice recursion is much more expensive than iteration.

2. Replace ioutil.ReadDir, which will both sort and stat files, so
unnecessarily slow.

3. Remove lots of filepath operations, including Cleans. All the Cleans
we called were redundant with filepath pkg which calls Clean
or a lot of its functions.

I've also taken the chance to refactor loadFiles into something
more readable IMO. This does add a bit of code dupping (specifically,
`shouldIgnore` and `meta.Err` checks), but the different handlings
are more explicit.