This gets rid of a few bottlenecks:
1. Recursion -> iteration. I'm not sure how much of a bottleneck this is,
but in practice recursion is much more expensive than iteration.

2. Replace ioutil.ReadDir, which will both sort and stat files, so
unnecessarily slow.

3. Remove lots of filepath operations, including Cleans. All the Cleans
we called were redundant with filepath pkg which calls Clean
or a lot of its functions.

As a side note, for future improvements on loadFiles, it may be worth
it to skip on some filepath functions. Per pprof analysis, 20% of CPU
time is still spent on Clean functions which are automatically called.

I've also taken the chance to refactor loadFiles into something
more readable IMO. This does add a bit of code dupping (specifically,
shouldIgnore and meta.Err checks), but the different handlings
are more explicit.

I've tried this on GCP e2-standard-8 machines and these changes make
Compute a Merkle Tree for an entire chrome build take about ~50% of
the time (~50s instead of ~100s), working in a single thread. Weirdly,
changing pd-standard vs pd-ssd doesn't seem to make a difference.