Simplify caching packages further (#306)
* Simplify caching packages further

Instead of using a global hashtable of caches with a single key, use
a global variable of the that cache.

Remove contents of "cache" package, while leaving the
"singleflightcache" subpackage.

Fix a bug: fmCache.Reset() cleared caches in all instances of fmCache
instead of just that one. Make the function global and clealry document
that it invalidates the cache globally.

Fix a couple of bugs around usage of atomic package along the way.