[cas] Add RPCConfig (#318)
* [cas] Add RPCConfig

First of all, this CL fixes a bug that max-bytes was confused with
max-count, in (*Client).Upload().

Stepping back, the reason is that the config names are long and
confusing. Earlier version of the code attempted to deal with by
partially following a rule that a per-RPC config starts with the RPC
name, but clearly this wasn't enough.

Introduce RPCConfig struct which contains RPC-specific configuration,
namely concurrency, max-size, max-count and timeout. Use the struct
as a type of two new fields: ClientConfig.FindMissingBlobs and
ClientConfig.BatchUpdateBlobs, which match RPC names exactly.

This is different from client package's approach, which is having a
map[string]RPCConfig. It was considered, but it causes non-type-safe
code, e.g. using string literals to refer to configuration of a
particular RPC. This could be mitigated by adding constant strings, but
this is no better than having strongly-typed fields instead of the map.

Also it seems like the main reason to use map in the client package is
to have defaults that apply to all RPCs, e.g. it is reasonable to have
1s as the default timeout. In constrast, the cas package defines more
than timeout, namely maxRequestSize and max-count, and those values
are rather different from RPC to RPC. So the wildcard approach used
in the client package provides less value.

Along all of this, this CL implements per-RPC timeouts, but it is a
rather small part of this CL.