[cas] Require PathSpec.Path to be absolute
A future CL with all PathSpec.Allowlist which is a list of
PathSpec.Path's descendants. It makes more sense for such paths to
be relative to PathSpec.Path than to CWD, but it would be confusing to
have/allow one struct (PathSpec) to contais paths relative to different
dirs.

Another reason to require PathSpec.Path to be absolute is because in
general CWD is arbitrary.