Preserving symlink or not can be configured from Command. (#343)
For some toolchains, a user need to run a wrapper to run a compiler.
Such a wrapper sometimes use readlink inside and resolving symlink
breaks such a script.  To make it work, we need to preserve symlinks.
fails to run.  The behavior is only needed for certain toolchains,
let me make it configurable via Command.InputSpec message.