Optionally process symlinks pointing outside exec root
Previously if `TreeSymlinkOpts.Preserved` was set to `true`, symlinks
pointing outside of the exec root would always be rejected. However, it
can be useful to preserve symlinks by default while replacing them with
the targeted file if the targeted file is outside of the exec root.

So add a new `TreeSymlinkOpts.MaterializeOutsideExecRoot` option that
enables this behavior if set to true.

See https://crbug.com/1216363 for context on why this is useful.