introduce the file system walker package (#440)
The package provides an API to walk the file system in depth-first style
traversal. This encapsulates IO and implementation naunces and exposes
a concise interface for clients.

The choice of having the client provide a set of callbacks instead of a
single shared one allows ensuring the client returns valid actions at
compile time. This is more robust than using implicit default actions or
propagating errors at runtime.

The DFS traversal and the deferring behaviour is leveraged in the CAS
upload implementation that will posted later.