fix regression in downloading out/err streams
The server may terminate the streaming request before sending available
bytes, which prevents the download fallback from getting them.

This change implements a resuming fallback to download the remainder of
the streams.

Additionally, the digests of the out and err streams will be printed in
the output of `execution_action` command to allow the user to access
them even if the remote execution fails, in which case they will not be
retrievable by `show_action`.