generic filter for walker and faster impath (#480)
I originally opted for a deterministic ID for the filter by using a regexp.
This turned out to be too restrictive and forces users to be creative with
regular expressions where simpler Go expressions would suffice.

The new implementation delegates the responsibility of using a unique
and deterministic filter ID to the user. The filter ID is required
for proper merkle tree processing in the casng uploader.

Additionally, a quick CPU profiling revealed a non-trival amount
of CPU time was spent on `filepath.Clean`, which is called by various
functions in the same package.
Building on the fact that paths constructed by the impath package are already
clean, I have updated the package to try to avoid unnecessary cleaning.

A subsequent quick profiling showed much less time spent on cleaning paths.
However, the cost might have been masked rather than eliminated, but
that is a question for antother day.