deadlock fix and cleanup
There are three deadlock fixes in this change. The first is described in code comments.

The second was caused by a regression in `excludeTag` function where it mutated
the underlying array that is shared with the caller. This caused a tag to be removed
and a tag to be duplicated, which eventually left some workers waiting indefinitely
for messages that were never delivered. It also caused other workers to have more
responses than expected which manifested as negative counts in the dispatcher.

The third was a potential deadlock I spotted while in `wait` method for
`pubsub`, which would hold a read lock until a done signal is sent,
potentially blocking any writes what should trigger said done signal.