casng, implement public upload methods (#483)
Upload is the main method used to upload a single request, which may
include a tree of files to upload.
The actual upload flow is defined in another file which will be
submitted later.

DigestTree skips uploading to the server and returns a merkle tree root
which can be useful for clients that wish to build upon that.

UploadTree covers the use case of having a list of inputs to upload
which requires replicating a filesystem tree rooted at the common
ancestor.
Using Upload with the root forces the user to craft an exclusion filter
that ensures only desired files are included. However, in build systems
the list of included files is readily available.
This method reconstructs the tree in memory without forcing the user to
recreate it on the filesystem (i.e. copy the files into an isoalted
root).