casng, implement stream pipe and digester worker (#485)
The stream pipe is used internally to pipe requests from public methods
to the shared internal processor, which uses a streaming interface.

The pipe control flow follows its input and output channels. Once the
input channel is closed, a control signal is sent to the processor which
is forwarded back to the pipe once the processor has processed all
requests origniating from that pipe.
Once the pipe receives the termination signal from the processor, it
closes its output channel allowing the public method to return
(batching) or close its own output channel.

The digestion worker manages file system walks and forwards generated
digests to the dispatcher which will be implemented in a subsequent
change. The dispatcher would be forwarding the blobs to gRPC workers
which would then publish the results back to the pipe.