casng, implement digestion methods
Each method accepts a filepath and produces a digested node, which
the worker then forwards to the dispatcher to be dispatched for uploading.

The file digestion method also handles acquiring and releasing IO tokens
to respect resource quotas. One particular nuance here is the handling
of tokens for large files. Such tokens are acquired in the digester here
but are not released until later in another worker. This leverages file
locality when processing many large files, such that the file is opened
once for both digestion and streaming to the remote server.

The tradeoff here is more susceptibility to bugs where the tokens are not
released correctly. For example, adding a return statement and
forgetting to update the position of the release statement.

To mitigate such risks, log messages are added to report on the total
number of tokens periodically (logBeat) and per request.