casng, implement the dispatcher (#487)
The dispatcher routes requests between three different components: the
digester, the gRPC workers, and the publishers.

The digester sends incoming blobs that need to be uploaded. The
dispatcher routes those to either the batching gRPC worker, which
handles small files, or to the streaming gRPC worker, which handles
large files.

The gRPC workers upload the blobs to the remote server and forward the
responses back to the dispatcher. The dispatcher then routes each
response to all relevant requesters (a response may be associated with
multiple requests).

The dispatcher is also responsible for routing control signals. It keeps
track of the number of blobs that were routed for each requester and
match that with the number of responses to the same requester. Once the
requester signals that it's done sending blobs, the dispatcher will
signal back once it's routed (drained) all remaining responses for that
requester. This ensures strong delivery guarantees and avoids losing any
message throughout the pipeline.