casng, implement gRPC batching worker
The gRPC batcher handles the batching API of the CAS. It unifies
duplicate blobs from multiple requesters into a single request to the
remote server, and returns the response to all related requesters.

The unification parameters are configurable and described in the
configs.go file as well as in the godocs of public methods.

Dispatching based on byte-size attempts to be as accurate as possible.
The bytes are padded by the overhead of serialized proto structures
rather than using an arbitrary approximation.

The retry loop is per batch, and only retries failed requests within a
batch.