casng, implement gRPC streaming worker (#492)
* casng, implement gRPC streaming worker

The gRPC streamer handles the byte stream API of the CAS. It unifies
duplicate blobs from multiple requesters while uploading the same blob
and returns the response to all releated requesters.

The unification is slightly different than the batching worker. Bundling
parameters from the gRPC config are not used. The streamer starts streaing
as soon as it receives a blob and it has available concurrent quota. This
means that unification only applies if an identical blob is received while
streaming. Otherwise, a streaming request will be initiated but the server
will respond with a cache hit.

The streamer also handles releasing IO holds for large files that
were acquired during digestion.

The retry loop is per chunk, not blob.

* casng, remove ctx field and minor cleanup