Fix contextmd.Infof (#502)
* Add KeepAlive support

KeepAlive functionality sends periodic pings to the server
over the transport layer.

* Change grpc_keepalive_permit_without_stream to false

There's no data indicating that the option provides any benefit;
however, having the option enabled  might introduce unnecessary
load on the GFEs.

* Fix contextmd.Infof

Fixes a bug in Infof that prevents the content of the context
to be appended to log messages.