casng, fix resolving remote paths
A few corner cases were not handeled properly.
The original one was not removing the leading slash after stripping
workingDir.
However, a more interesting one was that paths being outside workingDir,
but still under root, which could translate to partial replacements of
the workingDir path, which is a valid case, contrary to what I thought.
Reverted to using proper filepath APIs.