Fix stats in BatchDownloadBlobs call
Previously, BatchDownloadBlobs() wasn't properly recording the
compressed download size. To fix it, I've introduced a new public
interface called BatchDowloadBlobsWithStats(), which will return
statistics on the downloaded blob size in addition to the actual
content.

As part of this, I also realized that tests are non-existent for
compressed BatchDownloadBlobs() and added them.

In a subsequent change, I'll fix BatchUploadBlobs() as well.

TESTED: Unit test