fix preserving intermediate symlinks (#511)
If a filepath includes a symlink to a directory, that symlink is
converted to a directory in the computed Merkle tree. This is fine
unless symlink preservation is turned on, in which case any process that
expects a symlink might fail.

For example, if /a/b/c.go is loaded and b->bb, the Merkle tree will
only have b as a directory node and it will not include bb.
If /a/b and /a/b/c.go are loaded, the server will fail to write
both a directory and a symlink to the same path /a/b.
The symlink comes from loading /a/b and the directory comes from
loading /a/b/c.go.

The correct Merkle tree in this case should look like:

a
|_b->bb
|_bb
|_|_c.go

That is, any intermediate symlink to a directory appears in the tree
exactly once as a symlink. I.e. it never appears as a directory.

This change introduces additional IO overhead for each loaded file. The
overhead is proportional to the number of files, the depth of each
filepath, and the number of symlinks in each filepath.
Additionally, a map is introduced to deduplicate loaded files, which will
add memory and GC overhead proportional to the number of loaded files.
Both overheads are limited to requests with symlink preservation turned on.