fix preserving intermediate symlinks
This patch fixes an oversight in the previous PR #511 where an
intermediate symlink is processed just like an input which means all of
its tree is also included even though it is not part of the specified
input.

That not only inflated the Merkle tree, but also included symlinks to
targets outside the execution root which caused failures in builds that
do not specify the materialization option.