fix preserving intermediate symlinks
This patch handles virtual inputs. The code that handles virtual inputs
skips loadFiles entirely which effectively skips the previous patches.

This patch removes the redundant map by leveraging the existing one, fs.
The tradeoff is less allocation from the removed map, but more calls to
filepath.Rel and filepath.Join to get the key for the existing map.

Tested on a relevant action locally and confirmed it succeeds.
Also ran a full build with no issues.