Ignore setting PerRPCCreds as clientopts
When PerRPCCreds is used, we are already setting it as part of
DialParams, so we don't have to once again include it as part of
clientopts when creating the connection.
When included again as part of clientopts, it results in gRPC adding
duplicate Authorization header to the requests.

Issue: https://github.com/bazelbuild/reclient/issues/27 (thanks to
@MarshallOfSound for debugging this)

Tested: I build with this, ran a sample command and confirmed with
mitmproxy that authorization header wasn't repeated.