add dir content cache for ComputeMerkleTree
FMC (file metadata cache) helps avoid IO cost for digesting files.
When computing a merkle tree the directories still need to be listed
to construct the tree. Caching the listing helps avoid the cost of IO for that.

However, this might break some builds that have directories with varying
content throughout the build. For now, the cache is gated by a flag
to help guide a future change on how to filter out such directories
from the cache.