Fix GetTree fakes (#539)
`GetTree` was iterating over the `FileNode`s and trying to parse the blob as a `Directory`, it should only need to iterate over the `DirectoryNode`s. It was also adding the `Directory` to the `res`ult list when adding and removing the `Directory` from the queue, doubling the output.

The output now matches the RBE server, at least for a simple example.