Avoid racing in GetFlattenedOutputs()
GetFlattenedOutputs() takes the pointer of ec.resPb, and forward it to a long-running grpc call.

This will cause racing to happen in re-client unit test framework. See this one line change in re-client that face racing issue: tg/2184373

This change avoid the racing issue by make a deepcopy of resPb before forward it to grpc call.

Test: import to re-client and run  //internal/pkg/reproxy:reproxy_test with race detection.