Ignore directories without read permissions
Currently we check if we can read the metadata from the file/directory. Even if we can stat the file, it does not necessarily mean we have permissions to open the file/directory. This becomes a problem when we try to open the directory when computing the merkle tree and we don't have read permissions.