remove unmaintained errors pkg (#552)
The errors package was deprecated in 2021 https://github.com/pkg/errors

I converted wrapped errors to the standard method assuming the stack trace was not intended to be included in these calls.

I opted to removing the stack trace from the calls that explicitly included it in favour of proper error wrapping and propagation. I haven't needed to look at stack traces so far, but I'm not sure if others rely on it. This can be re-added later if needed.