Add flag to set symlink following behavior for uploading directory (#591)
We would like to be able to use upload_dir in a situation that makes heavy use of symlinks. However, the previous implementation did not respect symlinks and instead resulted in resolving the symlink to the actual path. This is problematic because it means that duplicated data was going to be recreated on a download.

This is solved by introducing a symlink_behavior argument that accepts unspecified (default), preserve, and resolve which map to the SymlinkBehaviorType enum. This allows the caller to specify how to handle symlinks and ensure that later downloads match exactly what the upload was. This also appears to reduce the data uploaded.