Update bazel version and fix race condition in test
To open this project with GoLand + bazel plugin, the min version of bazel we need is 6.0.0.

There is a race condition in the test that `batch_retries_test.go` cannot run together with the rest of the test files.

This CL fix add a lock to fix the race condition. add "exclusive" bazel tag to not run client_test with other bazel test targets, and
add an arg `-test.parallel=1` to make go run all the test case under client_test one by one.

Test: run bazelisk test //... --test_output=streamed --flaky_test_attempts=1 on my cloudtop
Bug: NA