Allow download outDir to be relative
Our previous change is blocking outDir to be . during CAS download.
Also there is a bug that getAbsPath() will not really return a abs
path if base is relative path, like foo/bar (it will return a
relative path based on foo/bar instead)

It is not necessay for blocking outDir to be ., we should allow
download to happen as long as the output files are not escaping
from outDir. This CL reslove the base (outDir) to an absolute path
then generate the full path for output files.